Fixed some compiler warnings.
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/900



git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@4401 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/test/intltest/regcoll.cpp
@@ -11,7 +11,7 @@
 #include "regcoll.h"
 #include "sfwdchit.h"
 
-#define ARRAY_LENGTH(array) (sizeof array / sizeof array[0])
+#define ARRAY_LENGTH(array) ((int32_t)(sizeof array / sizeof array[0]))
 
 CollationRegressionTest::CollationRegressionTest()
 {
